Warmups/MatryoshkaQR/requirements.txt
pillow==10.4.0
pyzbar==0.1.9
